nftables accelerated datapath (flowtables)
Allow users to select a subset of devices on the system to accelerate
the established Service connections via the kernel flowtables
infrastructure.

The user can pass a CEL expression to match on interface names or types,
the choice of CEL vs regex is that allow to provide much better
experience and extend the variables to other interface attributies.

The new options can only be used with kube-proxy nftables.