staticpolicy:smtalign: count for pre-allocated cpus for container
When CPUManager is configured with full-pcpus-only option,
it validates the number of available CPUs on the system.

There's a subtle bug in this check, where after kubelet restart
the containers already assigned with CPUs because
kubelet reads the checkpoint file and builds the container's
assignment map on startup.

We are not counting for these pre-allocated CPUs,
which could lead to false-positive, and prevent
the pod from being admitted.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>